requirements.txt
pymodbus < 3.0
paho-mqtt
ConfigArgParse
